Moved the GCT specific user update functions to the GCT context
