+++ lib/glimesh/accounts.ex
@@ -312,10 +312,6 @@ defmodule Glimesh.Accounts do
     User.profile_changeset(user, attrs)
   end
 
-  def change_user(user, attrs \\ %{}) do
-    User.gct_user_changeset(user, attrs)
-  end
-
   @doc """
   Updates the user password.
 
@@ -331,12 +327,6 @@ defmodule Glimesh.Accounts do
   def update_user_profile(user, attrs) do
     user
     |> User.profile_changeset(attrs)
-    |> Repo.update()
-  end
-
-  def update_user(user, attrs) do
-    user
-    |> User.gct_user_changeset(attrs)
     |> Repo.update()
   end
